move all forward CDI functions in a separate module
