+++ bcdi/preprocessing/__init__.py
@@ -7,11 +7,13 @@
 #       authors:
 #         Jerome Carnis, carnis_jerome@yahoo.fr
 """
-BCDI preprocessing.
+BCDI and forward CDI preprocessing.
 
 This package contains:
- - preprocessing_utils: functions related to data loading and preprocessing,
+ - preprocessing_utils: functions related to data loading and preprocessing for BCDI,
    before phase retrieval.
+ - tomography_utils: functions related to data loading and preprocessing for forward
+   CDI 3D datasets, before phase retrieval.
  - nxsReady, ReadNxs3: modules to load data at SIXS beamline
 
 """
